Catch expected pytest warning about invalid COT_URL
Running the suite of pytest tests yields a single warning:

```
tests/test_client_functions.py::test_protocol_factory_bad_url
  Warning: Invalid COT_URL=udp:localhost
```

Based on the context of the test, this warning shouldn't be flagged
as a warning at all. The warning is expected. This commit explicitly
catches the expected warning from pytest and marks the test as an
unconditional pass.